# finanzas-backend/src/main/resources/openapi/finanzas-api.yaml
openapi: 3.0.3
info:
  title: FinanzasApp API
  version: 2.0.0
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      summary: Registrar usuario
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Usuario creado
  /api/auth/login:
    post:
      summary: Iniciar sesion
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sesion iniciada
  /api/auth/google:
    post:
      summary: Completar Google OpenID Connect con Authorization Code + PKCE
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Sesion iniciada o cuenta vinculada
...
  /api/workspaces:
    get:
      summary: Listar workspaces del usuario
      responses:
        '200':
          description: Workspaces autorizados
    post:
      summary: Crear workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceCreateRequest'
      responses:
        '201':
          description: Workspace creado
  /api/workspaces/{workspaceId}:
    get:
      summary: Obtener workspace autorizado
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Workspace
...
  /api/workspaces/{workspaceId}/categories:
    get:
      summary: Listar categorias con filtros
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: type
          in: query
          schema:
            type: string
            enum: [INCOME, EXPENSE]
        - name: includeArchived
          in: query
          schema:
            type: boolean
            default: false
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Categorias
    post:
      summary: Crear categoria
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Categoria creada
  /api/workspaces/{workspaceId}/categories/{categoryId}:
    put:
      summary: Editar categoria
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '200':
          description: Categoria actualizada
    delete:
      summary: Archivar categoria
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Categoria archivada
  /api/workspaces/{workspaceId}/categories/{categoryId}/restore:
    post:
      summary: Restaurar categoria archivada conservando UUID
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Categoria restaurada
...
  /api/workspaces/{workspaceId}/budgets:
    get:
      summary: Listar presupuestos
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Presupuestos
    post:
      summary: Crear presupuesto
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetRequest'
      responses:
        '201':
          description: Presupuesto creado
  /api/workspaces/{workspaceId}/budgets/{budgetId}:
    put:
      summary: Editar presupuesto
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: budgetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetRequest'
      responses:
        '200':
          description: Presupuesto actualizado
    delete:
      summary: Eliminar presupuesto
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
        - name: budgetId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Presupuesto eliminado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    RegisterRequest:
      type: object
      required: [nombre, email, password]
      properties:
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        moneda:
          type: string
          minLength: 3
          maxLength: 3
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    GoogleAuthRequest:
      type: object
      required: [authorizationCode, codeVerifier, redirectUri]
      properties:
        authorizationCode:
          type: string
        codeVerifier:
          type: string
        redirectUri:
          type: string
...
    WorkspaceCreateRequest:
      type: object
      required: [nombre, tipo]
      properties:
        nombre:
          type: string
        tipo:
          type: string
          enum: [PERSONAL, HOUSEHOLD, BUSINESS]
...
    CategoryRequest:
      type: object
      required: [nombre, type]
      properties:
        nombre:
          type: string
        type:
          type: string
          enum: [INCOME, EXPENSE]
        icono:
          type: string
        color:
          type: string
...
    BudgetRequest:
      type: object
      required: [categoryId, amount, periodMonth]
      properties:
        categoryId:
          type: string
          format: uuid
        amount:
          type: number
          format: decimal
        periodMonth:
          type: string
          format: date
...